+++ src/Sastrawi/Stemmer/Stemmer.py
@@ -2,6 +2,7 @@ import re
 from Sastrawi.Stemmer.Context.Visitor.VisitorProvider import VisitorProvider
 from Sastrawi.Stemmer.Filter import TextNormalizer
 from Sastrawi.Stemmer.Context.Context import Context
+from cachetools import cached, LRUCache
 
 class Stemmer(object):
     """Indonesian Stemmer.
@@ -28,12 +29,23 @@ class Stemmer(object):
 
         return ' '.join(stems)
 
+    @cached(cache=LRUCache(maxsize=128))
     def stem_word(self, word):
         """Stem a word to its common stem form."""
         if self.is_plural(word):
             return self.stem_plural_word(word)
         else:
             return self.stem_singular_word(word)
+
+    # Stemming word in Tokens
+    # @author Mufid Jamaluddin <mufidjamaluddin@outlook.com>
+    def stem_tokens(self, tokens):
+        stemmed_tokens = []
+        for token in tokens:
+            if not token or token.strip() == '':
+                continue
+            stemmed_tokens.append(self.stem_word(token))
+        return stemmed_tokens
 
     def is_plural(self, word):
         #-ku|-mu|-nya
